(#102) Modularize login persistence helpers
